+++ lib/Point.h
@@ -38,11 +38,11 @@ public:
 	DPoint();
 	DPoint(double time, const QString &text);
 
-	double left();
-	double right();
+    double left() const override;
+    double right() const override;
 
-	bool precedes(Item *item);
-	bool follows(Item *item);
+    bool precedes(Item *item) const override;
+    bool follows(Item *item) const override;
 
 	double time();
 	void setTime(double);